Implemented noSchemeConcept and noTopConcepScheme ICV and improvements in danglingConcept.

+++ app/src/icv/danglingConcept/danglingConceptComponent.ts
@@ -16,7 +16,7 @@ import {SkosServices} from "../../services/skosServices";
 export class DanglingConceptComponent {
     
     private scheme: ARTURIResource;
-    private danglingConceptList: Array<ARTURIResource>;
+    private brokenConceptList: Array<ARTURIResource>;
     
     constructor(private icvService: IcvServices, private skosService: SkosServices, 
             private vbCtx: VocbenchCtx, private router: Router) {
@@ -29,21 +29,19 @@ export class DanglingConceptComponent {
     }
     
     /**
-     * Run the check and lists the dangling concepts
+     * Run the check
      */
     runIcv() {
-        //TODO when listDanglingConcepts will be refactored provide also scheme as parameter
+        //TODO check when service will be refactored
         document.getElementById("blockDivIcv").style.display = "block";
         this.icvService.listDanglingConcepts().subscribe(
             stResp => {
-                //TODO when the service will be refactored, revise this code
-                this.danglingConceptList = new Array();
+                this.brokenConceptList = new Array();
                 var recordColl = stResp.getElementsByTagName("record");
                 for (var i = 0; i < recordColl.length; i++) {
-                    //check needed untill listDanglingConcepts will be refactored
                     if (recordColl[i].getAttribute("scheme") == this.scheme.getURI()) {
                         var dc = new ARTURIResource(recordColl[i].getAttribute("concept"), recordColl[i].getAttribute("concept"), "concept"); 
-                        this.danglingConceptList.push(dc);       
+                        this.brokenConceptList.push(dc);       
                     }
                 }
             },
@@ -55,15 +53,16 @@ export class DanglingConceptComponent {
     }
     
     /**
-     * Fixes dangling concept by setting the concept as topConceptOf the current scheme, then update the danglingConceptList 
+     * Fixes concept by setting the concept as topConceptOf the current scheme 
      */
     setAsTopConcept(concept: ARTURIResource) {
         this.skosService.addTopConcept(concept, this.scheme).subscribe(
             data => {
                 //remove the concept from the danglingConceptList
-                for (var i = 0; i < this.danglingConceptList.length; i++) {
-                    if (this.danglingConceptList[i].getURI() == concept.getURI()) {
-                        this.danglingConceptList.splice(i, 1);
+                for (var i = 0; i < this.brokenConceptList.length; i++) {
+                    if (this.brokenConceptList[i].getURI() == concept.getURI()) {
+                        this.brokenConceptList.splice(i, 1);
+                        break;
                     }
                 }
             },
@@ -74,57 +73,60 @@ export class DanglingConceptComponent {
     }
     
     /**
-     * Fixes dangling concept by selecting a broader concept
-     */
-    selectBroader(concept: ARTURIResource) {
-        alert("Selecting broader for " + concept.getURI() + " (fix not yet available, opening concept tree for scheme "
-            + this.scheme.getURI() + "...)");
-        //TODO here I should open a modal to show the concept tree and select a concept
-    }
-    
-    /**
-     * Fixes dangling concept by removing the concept from the current scheme, then update the danglingConceptList 
-     */
-    removeFromScheme(concept: ARTURIResource) {
-        this.skosService.removeConceptFromScheme(concept, this.scheme).subscribe(
-            data => {
-                //remove the concept from the danglingConceptList
-                for (var i = 0; i < this.danglingConceptList.length; i++) {
-                    if (this.danglingConceptList[i].getURI() == concept.getURI()) {
-                        this.danglingConceptList.splice(i, 1);
-                    }
-                }
-            },
-            err => {
-                alert("Error: " + err);
-                console.error(err['stack']);
-            });
-    }
-    
-    /**
-     * Fixes all dangling concepts by setting them all as topConceptOf the current scheme
+     * Fixes all concepts by setting them all as topConceptOf the current scheme
      */
     setAllTopConcept() {
         //TODO this fix requires a new service server side that takes a list of concept and sets them as topConcept of a scheme
-        alert("Setting all concept as topConceptOf " + this.scheme.getURI() + " (quick fix not yet available)");
+        alert("Fix not yet available");
     }
     
     /**
-     * Fixes all dangling concepts by selecting a broader concept for them all 
+     * Fixes concept by selecting a broader concept
      */
-    selectBroaderForAll() {
-        //TODO this fix requires a new service server side that takes a list of concept and sets for them a broader concept
-        alert("Selecting a broader for all concept (quick fix not yet available, opening concept tree for scheme "
-            + this.scheme.getURI() + "...)");
+    selectBroader(concept: ARTURIResource) {
+        alert("Fix not yet available");
         //TODO here I should open a modal to show the concept tree and select a concept
     }
     
     /**
-     * Fixes all dangling concepts by removing them all from the current scheme 
+     * Fixes all concepts by selecting a broader concept for them all 
+     */
+    selectBroaderForAll() {
+        //TODO this fix requires a new service server side that takes a list of concept and sets for them a broader concept
+        alert("Fix not yet available");
+        //TODO here I should open a modal to show the concept tree and select a concept
+    }
+    
+    /**
+     * Fixes concept by removing the concept from the current scheme 
+     */
+    removeFromScheme(concept: ARTURIResource) {
+        if (confirm("Warning, if this concept has narrowers, removing the dangling concept from the scheme " +
+                "may generate other dangling concepts. Are you sure to proceed?")) {
+            this.skosService.removeConceptFromScheme(concept, this.scheme).subscribe(
+                data => {
+                    //remove the concept from the danglingConceptList
+                    for (var i = 0; i < this.brokenConceptList.length; i++) {
+                        if (this.brokenConceptList[i].getURI() == concept.getURI()) {
+                            this.brokenConceptList.splice(i, 1);
+                            break;
+                        }
+                    }
+                },
+                err => {
+                    alert("Error: " + err);
+                    console.error(err['stack']);
+                }
+            );
+        }
+    }
+    
+    /**
+     * Fixes concepts by removing them all from the current scheme 
      */
     removeAllFromScheme() {
         //TODO this fix requires a new service server side that takes a list of concept removes them from a scheme
-        alert("Removing all concept from " + this.scheme.getURI() + " (quick fix not yet available)");
+        alert("Fix not yet available");
     }
     
 }
